+++ job_templates/sag_np_cell_pipe/config_fed_client.conf
@@ -79,7 +79,7 @@
       args {
         mode = "PASSIVE"
         site_name = "{SITE_NAME}"
-        token = "pipe"
+        token = "{JOB_ID}"
         root_url = "{ROOT_URL}"
         secure_mode = false
         workspace_dir = "{WORKSPACE}"
